move adapter code to OperationPair
